Separate user-related backend code from file code
